Update Centos makefile and address issue #20
The Centos makefile was producing debug output, which was wrong.

The quote issue (#20) is reduced, but not defeated. The difficulty remains that to properly deal with the matter, ssc has to parse scripts and styles, which is not a small undertaking. Users can use the clunky CDATA syntax to manage things for now, if they wish.